requirements.txt
coverage==3.6
django-coverage
Pillow
